Improvements in KCss classes. Complex css (like margin-*) property support is added.
