--- karibu-dsl-v10/src/main/kotlin/com/github/vok/karibudsl/flow/KCssCustomeStyles.kt
@@ -1,5 +1,0 @@
-package com.github.vok.karibudsl.flow
-
-val cssBlueColorStyle: KCssStyleBlock = {
-    color = KCssColor.Blue.value
-}
