+++ karibu-dsl-v10/src/main/kotlin/com/github/vok/karibudsl/flow/cssstyledsl/KCssColor.kt
@@ -1,4 +1,4 @@
-package com.github.vok.karibudsl.flow
+package com.github.vok.karibudsl.flow.cssstyledsl
 
 enum class KCssColor(val value: String) {
     Blue("blue"),
