+++ karibu-dsl-v10/src/main/kotlin/com/github/vok/karibudsl/flow/cssstyledsl/KCssCustomeStyles.kt
@@ -1,0 +1,7 @@
+package com.github.vok.karibudsl.flow.cssstyledsl
+
+import com.github.vok.karibudsl.flow.cssstyledsl.properties.color
+
+val cssBlueColorStyle: KCssStyleBlock = {
+    color = KCssColor.Blue.value
+}
