+++ karibu-dsl-v10/src/main/kotlin/com/github/vok/karibudsl/flow/cssstyledsl/properties/KCssStyleAbstractPropertyA.kt
@@ -1,0 +1,5 @@
+package com.github.vok.karibudsl.flow.cssstyledsl.properties
+
+import com.vaadin.flow.dom.Style
+
+abstract class KCssStyleAbstractPropertyA(protected val style: Style)
